+++ server.js
@@ -1,7 +1,7 @@
 const express = require("express");
 const cors = require('cors');
 const mongoose = require("mongoose");
-const Router = require("./routes")
+const Router = require("./routes/routes")
 require('dotenv').config();
 
 const PORT = process.env.PORT || 3000;
